+++ lib/dart_faye.dart
@@ -14,11 +14,12 @@ export 'src/protocol/grammar.dart';
 export 'src/protocol/dispatcher.dart';
 export 'src/util/constants.dart';
 export 'src/util/namespace.dart';
+export 'src/extension.dart';
 
 /// Main Faye library class containing constants and utility methods
 class Faye {
   /// Current version of the Dart Faye implementation
-  static const String version = '1.1.0';
+  static const String version = '1.2.0';
   
   /// Bayeux protocol version
   static const String bayeuxVersion = '1.0';